requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-spectacular==0.26.5
openpyxl==3.1.2
streamlit==1.28.1
requests==2.31.0
pandas==2.1.3
plotly==5.17.0
psycopg2-binary==2.9.10
python-decouple
